[JENKINS-39057] better storyboards for testing font changes
